fix: xbi codec was reading the wrong size for asset id

+++ sabi/src/lib.rs
@@ -8,7 +8,7 @@ use std::marker::PhantomData;
 /// Global XBI Types.
 pub type Data = Vec<u8>;
 /// A representation of an Asset Id, this is utilised for xbi instructions relating to multiple assets
-pub type AssetId = u64; // Could also be xcm::MultiAsset
+pub type AssetId = u32; // Could also be xcm::MultiAsset
 pub type Gas = u64;
 pub type AccountId32 = sp_runtime::AccountId32;
 pub type AccountId20 = sp_core::H160; // Could also take it from MultiLocation::Junction::AccountKey20 { network: NetworkId, key: [u8; 20] },
